% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_path.R
\name{ATTbrute}
\alias{ATTbrute}
\title{Find modulus using CVX}
\usage{
ATTbrute(delta2, D0)
}
\arguments{
\item{delta2}{square of delta for modulus}

\item{D0}{matrix of distances with dimension \code{[n1 n0]} between untreated
and treated units}
}
\description{
Find modulus using CVX
}
\keyword{internal}
