% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_path.R
\name{ATTcheck}
\alias{ATTcheck}
\title{Check solution validity using CVX}
\usage{
ATTcheck(m, r, mu, D0)
}
\arguments{
\item{m}{values for untreated observations}

\item{r}{values for treated observations}

\item{D0}{matrix of distances with dimension \code{[n1 n0]} between untreated
and treated units, where \code{n0} is the number of untreated units and
\code{n1} is the number of treated units}
}
\description{
Check solution validity using CVX
}
\keyword{internal}
