% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTEstimatePath}
\alias{ATTEstimatePath}
\title{build optimal estimator given a solution path}
\usage{
ATTEstimatePath(res, d, y, C = 1, sigma2, alpha = 0.05, beta = 0.8)
}
\arguments{
\item{res}{output of \code{ATTh}}

\item{d}{vector of treatment indicators}

\item{y}{outcome vector}

\item{C}{smoothness constant}

\item{sigma2}{vector of variances}

\item{alpha}{CI coverage}

\item{beta}{quantile of excess length}
}
\description{
build optimal estimator given a solution path
}
