% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_path.R
\name{ATTh}
\alias{ATTh}
\title{Homotopy for average treatment effect for the treated}
\usage{
ATTh(D0, maxiter = 50, check = FALSE, tol = .Machine$double.eps * n0 * n1)
}
\arguments{
\item{D0}{matrix of distances with dimension \code{[n1 n0]} between untreated
and treated units}

\item{maxiter}{maximum number of steps in the homotopy. If the homotopy has
less steps than \code{maxiter}, returns the whole solution path.}

\item{check}{check at each step that solution matches
\code{\link[CVXR]{CVXR}} (generic convex optimizer) solution.}

\item{tol}{numerical tolerance for rounding error when finding the nearest
neighbor. All observations with effective distance withing \code{tol} of
the closest are considered to be active.}

\item{s}{Set of state variables at which to start the homotopy. If not
    provided, the homotopy is started at the beginning. The state variables
    are as follows:
\describe{

  \item{m0}{A vector of length \code{n0} of corresponding to \eqn{m}}
  \item{r0}{A vector of length \code{n1} of corresponding to \eqn{r}}
  \item{mu}{A scalar corresponding to \eqn{\mu}}
  \item{D}{A matrix of effective distances with dimension \code{[n1 n0]}}
  \item{Lam}{A sparse matrix of Lagrange multipliers with dimension \code{[n1 n0]}}
  \item{N0}{A sparse matrix of nearest neighbors with dimension \code{[n1 n0]}}

}}
}
\value{
A list with two elements:

\describe{

  \item{res}{A matrix with rows corresponding to steps in the homotopy, so
  that the maximum number of rows is \code{maxiter}, and columns
  corresponding to \eqn{\delta}, \eqn{m}, \eqn{r}, \eqn{\mu}, and \code{drop} }

  \item{s}{List of state variables at the last step with the same structure
  as the input \code{s}}

}
}
\description{
Calculates optimal weights \eqn{m} and \eqn{r} as a function of \eqn{\delta},
or equivalently \eqn{mu}.
}
\examples{
x0 <- c(0, 1, 2, 3)
x1 <- c(1, 4, 5)
D0 <- distMat(x0, x1)
ATTh(D0)
}
