% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTmatch}
\alias{ATTmatch}
\title{Matching estimator for the ATT}
\usage{
ATTmatch(M, y, d, D0, sigma2)
}
\arguments{
\item{M}{number of matches}

\item{y}{outcome vector}

\item{d}{vector of treatment indicators}

\item{D0}{matrix of distances with dimension \code{[n1 n0]} between untreated
and treated units, where \code{n0} is number of untreated units and
\code{n1} is number of treated units}

\item{sigma2}{vector of variances}
}
\description{
Matching estimator for the ATT
}
