% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTOptPath}
\alias{ATTOptPath}
\title{build optimal estimator given a solution path}
\usage{
ATTOptPath(res, y, d)
}
\arguments{
\item{res}{The \code{res} element of the output of \code{ATTh} on which to
base the estimate}

\item{y}{outcome vector with length \code{n}}

\item{d}{Vector of treatment indicators with length \code{n}}

\item{alpha}{Level of confidence interval, \code{1-alpha}.}

\item{beta}{The quantile \code{beta} of excecss length for
determining performance of one-sided CIs.}

\item{sigma2}{vector of variance estimates with length \code{n}}

\item{C}{Lipschitz smoothness constant}
}
\description{
build optimal estimator given a solution path
}
