% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTpath}
\alias{ATTpath}
\title{Estimator along the optimal solution path}
\usage{
ATTpath(res, d, y, C = 1, sigma2, alpha = 0.05)
}
\arguments{
\item{res}{output of \code{ATTh}}

\item{d}{vector of treatment indicators}

\item{y}{outcome vector}

\item{C}{smoothness constant}

\item{sigma2}{vector of variances}

\item{alpha}{CI coverage}
}
\description{
Estimator along the optimal solution path
}
