% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_path.R
\name{ATTstep}
\alias{ATTstep}
\title{Calculate next step in ATT homotopy}
\usage{
ATTstep(s, tol = .Machine$double.eps * n0 * n1)
}
\arguments{
\item{s}{list of state variables}
}
\value{
Updated list \code{s} of state variables
}
\description{
Calculate next step in ATT homotopy
}
\keyword{internal}
