% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{AUObrute}
\alias{AUObrute}
\title{Find modulus using CVX}
\usage{
AUObrute(delta, D0)
}
\arguments{
\item{delta}{delta for modulus}

\item{D0}{matrix of distances with dimension \code{[n1 n0]} between untreated
and treated units}
}
\description{
Find modulus using CVX
}
