% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{AUOcheck}
\alias{AUOcheck}
\title{Check solution validity}
\usage{
AUOcheck(m, r, D0)
}
\arguments{
\item{m}{values for untreated observations}

\item{r}{values for treated observations}

\item{D0}{matrix of distances with dimension \code{[n1 n0]} between untreated
and treated units}
}
\description{
Check solution validity
}
