% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{AUOh}
\alias{AUOh}
\title{Average untreated effect homotopy}
\usage{
AUOh(D0, maxiter = 50, check = FALSE)
}
\arguments{
\item{D0}{matrix of distances with dimension \code{[n1 n0]} between untreated
and treated units}

\item{maxiter}{maximum number of steps in the homotopy}

\item{check}{check solution using CVX at each step}
}
\value{
Returns solution path
}
\description{
Average untreated effect homotopy
}
