% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{AUOstep}
\alias{AUOstep}
\title{Calculate next step in AUO homotopy}
\usage{
AUOstep(s, A)
}
\arguments{
\item{s}{list of state variables}
}
\description{
Calculate next step in AUO homotopy
}
\keyword{internal}
