% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv}
\alias{cv}
\title{Critical values for CIs around a biased Gaussian estimator.}
\usage{
cv(B, alpha = 0.05)
}
\arguments{
\item{B}{Maximum bias, a non-negative vector.}

\item{alpha}{Determines CI level, \code{1-alpha}. Needs to be between 0 and
1. Can be a vector of values.}
}
\value{
Critical value
}
\description{
Computes the critical value \eqn{cv_{1-alpha}(B)} that is needed to make the
confidence interval \eqn{X\pm cv_{1-alpha}(B)} have coverage \eqn{1-alpha} if
\eqn{X} is normally distributed with variance one and maximum bias at most
\eqn{B}.
}
\examples{
# 90\% critical value:
cv(B = 1, alpha = 0.1)
# 95\% critical values
cv(B = c(0, 1, 3), alpha = 0.05)
# 96 and 90\% critical values
cv(B = 1, alpha = c(0.05, 0.1))
}
