% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnm.R
\name{CVb}
\alias{CVb}
\title{Critical values for CIs based on a biased Gaussian estimator.}
\usage{
CVb(B, alpha = 0.05)
}
\arguments{
\item{B}{Maximum bias, a non-negative number.}

\item{alpha}{Determines CI level, \code{1-alpha}. Needs to be between 0 and
1. Can be a vector of values.}
}
\value{
Critical value
}
\description{
Computes the critical value \eqn{cv_{1-alpha}(B)} that is needed to make the
confidence interval \eqn{X\pm cv} have coverage \eqn{1-alpha} if \eqn{X} is
normally distributed with variance one and maximum bias at most \eqn{B}.
}
\examples{
# 90\% critical value:
CVb(B = 1, alpha = 0.1)
CVb(B = 1, alpha = c(0.05, 0.1))
}
