% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{NSW}
\alias{NSW}
\title{Dataset from Dehejia and Wahba (1999)}
\format{A data frame with 2,675 observations (2,490 controls, and 185
    treated) and 11 variables.

\describe{
\item{treated}{Treatment indicator}
\item{age}{}
\item{education}{}
\item{black}{}
\item{hispanic}{}
\item{married}{}
\item{re74}{}
\item{re75}{}
\item{re78}{}
\item{ue74}{Indicator for zero earnings in 1974}
\item{ue75}{Indicator for zero earnings in 1975}
}}
\source{
Rajeev Dehejia's website, \url{http://users.nber.org/~rdehejia/nswdata2.html}
}
\usage{
NSW
}
\description{
Subset of National Supported Work data from Dehejia and Wahba (1999).
}
\references{
{

\cite{Dehejia, R., and Wahba, S. (1999), "Causal Effects in Nonexperimental
Studies: Reevaluating the Evaluation of Training Programs," Journal of the
American Statistical Association, 94, 1053-1062.}

}
}
\keyword{datasets}
