% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{NSWexper}
\alias{NSWexper}
\title{Experimental dataset from Dehejia and Wahba (1999)}
\format{A data frame with 445 observations (185 treated anbd 260 controls)
    and 11 variables.

\describe{
\item{treated}{Treatment indicator}
\item{age}{Age in years}
\item{education}{Year of education}
\item{black}{Indicator for black}
\item{hispanic}{Indicator for Hispanic}
\item{married}{Indicator for married}
\item{re74}{Earnings in 1974 (in thousands of dollars)}
\item{re75}{Earnings in 1975 (in thousands of dollars)}
\item{re78}{Earnings in 1978 (in thousands of dollars)}
\item{ue74}{Indicator for zero earnings in 1974}
\item{ue75}{Indicator for zero earnings in 1975}
}}
\source{
Rajeev Dehejia's website, \url{http://users.nber.org/~rdehejia/nswdata2.html}
}
\usage{
NSWexper
}
\description{
National Supported Work data from Dehejia and Wahba (1999).
}
\references{
{

\cite{Dehejia, R., and Wahba, S. (1999), "Causal Effects in Nonexperimental
Studies: Reevaluating the Evaluation of Training Programs," Journal of the
American Statistical Association, 94, 1053-1062.}

}
}
\keyword{datasets}
