% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{nswhomotopy}
\alias{nswhomotopy}
\title{Homotopy for the data from Dehejia and Wahba (1999)}
\format{First 1,500 steps of the homotopy}
\usage{
nswhomotopy
}
\description{
Optimal solution path precomputed using the homotopy algorithm in this
package. It can be recomputed using the \code{ATTh} function, but takes a few
minutes
}
\keyword{datasets}
