% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{UpdatePath}
\alias{UpdatePath}
\title{update estimation path with new C or new variance}
\usage{
UpdatePath(ep, resw, C, sigma2, alpha, beta)
}
\description{
update estimation path with new C or new variance
}
\keyword{internal}
