% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTEffBounds}
\alias{ATTEffBounds}
\title{Efficiency bounds for confidence intervals}
\usage{
ATTEffBounds(h, d, sigma2, C = 1, beta = 0.8, alpha = 0.05)
}
\arguments{
\item{h}{The output of \code{ATTh}.}

\item{d}{Vector of treatment indicators with length \code{n}}

\item{sigma2}{Estimate of the conditional variance of the outcome, used to
optimize the number of matches.}

\item{C}{Lipschitz smoothness constant}

\item{beta}{The quantile \code{beta} of excess length for determining
performance of one-sided CIs.}

\item{alpha}{Level of confidence interval, \code{1-alpha}.}
}
\description{
Computes the asymptotic efficiency of two-sided fixed-length confidence
intervals at smooth functions, as well as the efficiency of one-sided
confidence intervals that optimize a given \code{beta} quantile of excess
length, using the formula described in Appendix A of Armstrong and Kolesár
(2018)
}
\references{
\cite{Armstrong, T. B., and M. Kolesár (2018): Finite-Sample
    Optimal Estimation and Inference on Average Treatment Effects Under
    Unconfoundedness, Unpublished manuscript}
}
