% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_path.R
\name{ATTh}
\alias{ATTh}
\title{Homotopy for average treatment effect for the treated}
\usage{
ATTh(
  D0,
  h,
  maxiter = 50,
  check = FALSE,
  tol = .Machine$double.eps * ncol(D0) * nrow(D0)
)
}
\arguments{
\item{D0}{matrix of distances with dimension \code{[n1 n0]} between untreated
and treated units, where \code{n0} is the number of untreated units and
\code{n1} is the number of treated units}

\item{h}{Optionally, supply previous output of \code{ATTh}. If not provided,
the homotopy is started at the beginning. If provided, it starts at the
step where the previous call to \code{ATTh} ended.}

\item{maxiter}{maximum number of steps in the homotopy. If the homotopy has
less steps than \code{maxiter}, returns the whole solution path.}

\item{check}{check at each step that the solution matches that in the
\code{\link[CVXR]{CVXR-package}} (generic convex optimizer package).}

\item{tol}{numerical tolerance for rounding error when finding the nearest
neighbor. All observations with effective distance within \code{tol} of
the closest are considered to be active.}
}
\value{
A list with two elements:

\describe{

  \item{res}{A matrix with rows corresponding to steps in the homotopy, so
  that the maximum number of rows is \code{maxiter} (if homotopy started at
  the beginning), and columns corresponding to \eqn{\delta}, \eqn{m},
  \eqn{r}, \eqn{\mu}, and \code{drop}, an indicator if an observations has
  been dropped from an active set, or added }

  \item{m0}{A vector of length \code{n0} of corresponding to \eqn{m} at the
  last step.}

  \item{r0}{A vector of length \code{n1} of corresponding to \eqn{r} at the
  last step.}

  \item{mu}{A scalar corresponding to \eqn{\mu} at the last step.}

  \item{D}{A matrix of effective distances with dimension \code{[n1 n0]} at
  the last step.}

  \item{Lam}{A sparse matrix of Lagrange multipliers with dimension
            \code{[n1 n0]} at the last step.}

  \item{N0}{A sparse matrix of nearest neighbors with dimension
            \code{[n1 n0]} at the last step.}
}
}
\description{
Calculates optimal weights \eqn{m} and \eqn{r} for the control and treated
observations as a function of \eqn{\delta}, or equivalently \eqn{\mu}, using
the algorithm described in the appendix to Armstrong and Kolesár (2018)
}
\examples{
x0 <- c(0, 1, 2, 3)
x1 <- c(1, 4, 5)
d <- c(rep(FALSE, length(x0)), rep(TRUE, length(x1)))
D0 <- distMat(c(x0, x1), d=d)
## Compute first three steps
h <- ATTh(D0, maxiter=3)
## Compute the remaining steps, checking them against CVX solution
h2 <- ATTh(D0, h=h, check=TRUE)
## Check
x0 <- c(1,  1,  3,  3,  4,  6, 17, 17, 21, 22, 23, 23, 26, 30, 36, 37,
        47, 53, 58, 61)
x1 <- c(9, 15, 26, 27, 28, 31, 32, 40, 51, 52, 56)
CVXR::installed_solvers()
d <- c(rep(FALSE, length(x0)), rep(TRUE, length(x1)))
D0 <- distMat(c(x0, x1), d=d)
ATTh(D0, check=TRUE)
## Check 2
dt <- NSWexper[c(1:15, 431:445), ]
X <- as.matrix(dt[, 2:10])
d <- dt$treated
D0 <- distMat(X, diag(c(0.15, 0.6, 2.5, 2.5, 2.5, 0.5, 0.5, 0.1, 0.1)),
                 method="manhattan", d)
sigma2 <- 40
h <- ATTh(D0, maxiter=300)
h$res
eb <- ATTEffBounds(h, d, sigma2, C=1)
}
\references{
\cite{Armstrong, T. B., and M. Kolesár (2018): Finite-Sample
    Optimal Estimation and Inference on Average Treatment Effects Under
    Unconfoundedness, Unpublished manuscript}
}
