% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTMatchEstimate}
\alias{ATTMatchEstimate}
\title{Inference on the CATT using the matching estimator}
\usage{
ATTMatchEstimate(mp, sigma2, C = 1, sigma2final = sigma2,
  alpha = 0.05, beta = 0.8, opt.criterion = "RMSE")
}
\arguments{
\item{mp}{Output of \code{ATTMatchPath}}

\item{sigma2}{vector of variance estimates with length \code{n}. If
\code{sigma2} is a scalar, it is assumed that the variance is
homoskedastic.}

\item{C}{Lipschitz smoothness constant}

\item{sigma2final}{vector of variance estimates with length{n} for
determining standard error of the optimal estimators. In contrast,
\code{sigma2} is used only for determining the optimal tuning parameter.}

\item{alpha}{Level of confidence interval, \code{1-alpha}.}

\item{beta}{The quantile \code{beta} of excecss length for determining
performance of one-sided CIs.}

\item{opt.criterion}{One of \code{"RMSE"} (root mean squared error),
\code{"OCI"} (one-sided confidence intervals), \code{"FLCI"}
(fixed-length two-sided confidence intervals)}
}
\description{
Computes matching estimator and confidence intervals (CIs) for the CATT. If
\code{ATTMatchPath} used a single \code{M}, the estimator and CIs are based
on a matching estimator with this number of matches. Otherwise, optimize the
numer of matches from the set in \code{M} according to \code{opt.criterion}.
}
\examples{
Ahalf <- diag(c(0.15, 0.6, 2.5, 2.5, 2.5, 0.5, 0.5, 0.1, 0.1))
D0 <- distMat(NSWexper[, 2:10], Ahalf, method="manhattan", NSWexper$treated)
mp <- ATTMatchPath(NSWexper$re78, NSWexper$treated, D0, M=c(1, 2), tol=1e-12)
## Distance matrix for variance estimation
DM <- distMat(NSWexper[, 2:10], Ahalf, method="manhattan")
sigma2 <- nnvar(DM, NSWexper$treated, NSWexper$re78, J=3)
## Estimator and CI based on a single match
ATTMatchEstimate(mp, mean(sigma2), C=1, sigma2final=sigma2)
}
