% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTMatchEstimate}
\alias{ATTMatchEstimate}
\title{build optimal estimator given a solution path}
\usage{
ATTMatchEstimate(mp, sigma2, C = 1, sigma2final = sigma2,
  alpha = 0.05, beta = 0.8, opt.criterion = "RMSE")
}
\arguments{
\item{mp}{Output of \code{ATTMatchPath}}

\item{sigma2}{vector of variance estimates with length \code{n}}

\item{C}{Lipschitz smoothness constant}

\item{sigma2final}{vector of variance estimates with length{n} for
determining standard error of the optimal estimators. In contrast,
\code{sigma2} is used only for determining the optimal tuning parameter.}

\item{alpha}{Level of confidence interval, \code{1-alpha}.}

\item{beta}{The quantile \code{beta} of excecss length for
determining performance of one-sided CIs.}

\item{opt.criterion}{One of \code{"RMSE"}, \code{"OCI"}, \code{"FLCI"}}

\item{y}{outcome vector with length \code{n}}

\item{d}{Vector of treatment indicators with length \code{n}}

\item{D0}{matrix of distances with dimension \code{[n1 n0]} between untreated
and treated units, where \code{n0} is number of untreated units and
\code{n1} is number of treated units}

\item{Mrange}{Range of Ms}
}
\description{
build optimal estimator given a solution path
}
