% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_match.R
\name{ATTMatchEstimate}
\alias{ATTMatchEstimate}
\title{Inference on the CATT using the matching estimator}
\usage{
ATTMatchEstimate(
  mp,
  sigma2,
  C = 1,
  sigma2final = sigma2,
  alpha = 0.05,
  beta = 0.8,
  opt.criterion = "RMSE"
)
}
\arguments{
\item{mp}{Output of \code{ATTMatchPath}}

\item{sigma2}{Estimate of the conditional variance of the outcome, used to
optimize the number of matches.}

\item{C}{Lipschitz smoothness constant}

\item{sigma2final}{vector of variance estimates with length \code{n} for
determining the standard error of the optimal estimator. In contrast,
\code{sigma2} is used only for determining the optimal tuning parameter.}

\item{alpha}{Level of confidence interval, \code{1-alpha}.}

\item{beta}{The quantile \code{beta} of excess length for determining
performance of one-sided CIs.}

\item{opt.criterion}{One of \code{"RMSE"} (root mean squared error),
\code{"OCI"} (one-sided confidence intervals), \code{"FLCI"}
(fixed-length two-sided confidence intervals)}
}
\value{
Returns an object of class \code{"ATTEstimate"}. An object of class
    \code{"ATTEstimate"} is a list containing the following components:
\describe{
\item{e}{Data frame with columns TODO}
\item{k}{weights TODO}
}
}
\description{
Computes matching estimator and confidence intervals (CIs) for the CATT. If
\code{ATTMatchPath} used a single \code{M}, the estimator and CIs are based
on a matching estimator with this number of matches. Otherwise, optimize the
number of matches from the set in \code{M} according to \code{opt.criterion}.
}
\examples{
Ahalf <- diag(c(0.15, 0.6, 2.5, 2.5, 2.5, 0.5, 0.5, 0.1, 0.1))
D0 <- distMat(NSWexper[, 2:10], Ahalf, method="manhattan", NSWexper$treated)
mp <- ATTMatchPath(NSWexper$re78, NSWexper$treated, D0, M=c(1, 2), tol=1e-12)
## Distance matrix for variance estimation
DM <- distMat(NSWexper[, 2:10], Ahalf, method="manhattan")
sigma2 <- nnvar(DM, NSWexper$treated, NSWexper$re78, J=3)
## Estimator based on a single match is better than with 2 matches for RMSE
ATTMatchEstimate(mp, mean(sigma2), C=1, sigma2final=sigma2)
}
