% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTMatchPath}
\alias{ATTMatchPath}
\title{Matching estimator for the ATT}
\usage{
ATTMatchPath(y, d, sigma2, D0, C, M, alpha = 0.05, beta = 0.8)
}
\arguments{
\item{y}{outcome vector with length \code{n}}

\item{d}{Vector of treatment indicators with length \code{n}}

\item{sigma2}{vector of variance estimates with length \code{n}}

\item{D0}{matrix of distances with dimension \code{[n1 n0]} between untreated
and treated units, where \code{n0} is number of untreated units and
\code{n1} is number of treated units}

\item{C}{Lipschitz smoothness constant}

\item{M}{number of matches}

\item{alpha}{Level of confidence interval, \code{1-alpha}.}

\item{beta}{The quantile \code{beta} of excecss length for
determining performance of one-sided CIs.}
}
\description{
Matching estimator for the ATT
}
