% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTOptEstimate}
\alias{ATTOptEstimate}
\title{Optimal estimation and inference for the CATT and the PATT}
\usage{
ATTOptEstimate(
  y,
  d,
  D0,
  path,
  C = 1,
  opt.criterion = "RMSE",
  sigma2init,
  sigma2,
  mvar,
  DM,
  alpha = 0.05,
  beta = 0.8,
  J = 3,
  extrasteps = 50
)
}
\arguments{
\item{y}{outcome vector with length \code{n}.}

\item{d}{vector of treatment indicators with length \code{n}.}

\item{D0}{matrix of distances with dimension \code{n1} by \code{n0} between
untreated and treated units, where \code{n0} is the number of untreated
units and \code{n1} is the number of treated units.}

\item{path}{Output of \code{ATTOptPath}. If this parameter is omitted, and
the parameters \code{y}, \code{d}, and \code{D0} are supplied,
\code{ATTOptEstimate} computes the path internally. Conversely, if
\code{path} is supplied, the parameters \code{y}, \code{d}, and \code{D0}
are ignored.}

\item{C}{Lipschitz smoothness constant}

\item{opt.criterion}{criterion to optimize. One of \code{"RMSE"} (root mean
squared error), \code{"OCI"} (one-sided confidence intervals),
\code{"FLCI"} (fixed-length two-sided confidence intervals)}

\item{sigma2init}{estimate of the conditional variance of the outcome, used
to choose the optimal smoothing parameter \eqn{\delta}{delta}. If not
supplied, use homoskedastic variance estimate based on a nearest neighbor
variance estimator.}

\item{sigma2}{vector of variance estimates with length \code{n} for
determining the conditional standard error of the optimal estimator. In
contrast, \code{sigma2init} is used only for determining the optimal
tuning parameter. If not supplied, use the nearest neighbor variance
estimator.}

\item{mvar}{Marginal variance estimate (variance of the CATT) used to
construct CIs for the PATT. If not supplied use the matching estimator of
Abadie and Imbens (2006).}

\item{DM}{distance matrix with dimension \code{n} by \code{n} to determine
nearest neighbors when when estimating \code{sigma2init} and
\code{sigma2}.}

\item{alpha}{determines confidence level, \code{1-alpha}.}

\item{beta}{quantile \code{beta} of excess length for determining performance
of one-sided CIs.}

\item{J}{number of nearest neighbors to use when estimating \code{sigma2init}
and \code{sigma2}.}

\item{extrasteps}{If the optimal smoothing parameter \eqn{\delta}{delta} is
attained at the end of the solution path \code{path}, compute additional
\code{extrasteps} steps in the solution path and estimate it again. If
\code{extrasteps==0}, then do not compute extra steps.}
}
\value{
Returns an object of class \code{"ATTEstimate"}. An object of class
    \code{"ATTEstimate"} is a list containing the following components:
    \describe{

    \item{e}{Vector with elements \code{"att"} (value of ATT estimate),
    \code{"maxbias"} (worst-case bias), \code{"delta"} (tuning parameter
    \eqn{\delta}{delta}), "omega" (\eqn{\omega(\delta)}{omega(delta)})
    \code{"lindw"} (maximal Lindeberg weight \eqn{Lind(k)}), \code{"sd"},
    \code{"hl"}, \code{"lower"}, \code{"upper"}, \code{"maxel"},
    \code{"rmse"} (standard deviation, half-length of two-sided CI, lower and
    upper endpoint of one-sided CIs, worst-case excess length of one-sided CI
    at quantile \code{beta} and RMSE of estimator, assuming conditional
    variance equals \code{sigma2inint}) \code{"rsd"}, \code{"rlower"},
    \code{"rupper"}, \code{"rhl"}, \code{"rrmse"}, \code{"rmaxel"} (same
    quantities, but calculated using \code{sigma2}), \code{"usd"},
    \code{"ulower"}, \code{"uupper"}, \code{"uhl"} (standard deviation,
    endpoints for one-sided CIs and half-length of two-sided CI for PATE),
    and \code{"C"} (Value of Lipschitz constant \code{C})}

    \item{k}{Vector of optimal weights \eqn{k(x_i, d_i)}}

    \item{res}{A vector with elements corresponding to the value of the state
    variables \eqn{\delta}, \eqn{m}, \eqn{r}, \eqn{\mu}, and \code{drop} of
    the solution path at the optimal value of \eqn{\delta}{delta}}

    \item{path}{The solution path, output of \code{ATTOptPath}.}

    }
}
\description{
Computes the estimator and confidence intervals (CIs) for the conditional
average treatment effect for the treated (CATT) and population average
treatment effect for the treated (PATT). The tuning parameter
\eqn{\delta}{delta} is chosen to optimize the criterion specified by
\code{opt.criterion}.
}
\examples{
## Use NSW experimental subsample with 25 treated and untreated units
dt <- NSWexper[c(1:25, 421:445), ]
Ahalf <- diag(c(0.15, 0.6, 2.5, 2.5, 2.5, 0.5, 0.5, 0.1, 0.1))
D0 <- distMat(dt[, 2:10], Ahalf, method="manhattan", dt$treated)
## Distance matrix for variance estimation
DM <- distMat(dt[, 2:10], Ahalf, method="manhattan")
c1 <- ATTOptEstimate(y=dt$re78, d=dt$treated, D0=D0, C=1, DM=DM,
                     opt.criterion="RMSE")
## Re-use the solution path already computed
c2 <- ATTOptEstimate(path=c1$path, C=1, DM=DM, opt.criterion="FLCI")
}
\references{
\cite{Armstrong, T. B., and M. Kolesár (2020): Finite-Sample
    Optimal Estimation and Inference on Average Treatment Effects Under
    Unconfoundedness, \url{https://arxiv.org/abs/1712.04594}}
}
