% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTOptEstimate}
\alias{ATTOptEstimate}
\title{Optimal estimation and inference for the CATT}
\usage{
ATTOptEstimate(op, sigma2, C = 1, sigma2final = sigma2, alpha = 0.05,
  beta = 0.8, opt.criterion = "RMSE")
}
\arguments{
\item{op}{Output of \code{ATTOptPath}.}

\item{sigma2}{vector of variance estimates with length \code{n}. If
\code{sigma2} is a scalar, it is assumed that the variance is
homoskedastic.}

\item{C}{Lipschitz smoothness constant}

\item{sigma2final}{vector of variance estimates with length{n} for
determining standard error of the optimal estimators. In contrast,
\code{sigma2} is used only for determining the optimal tuning parameter.}

\item{alpha}{Level of confidence interval, \code{1-alpha}.}

\item{beta}{The quantile \code{beta} of excecss length for determining
performance of one-sided CIs.}

\item{opt.criterion}{One of \code{"RMSE"} (root mean squared error),
\code{"OCI"} (one-sided confidence intervals), \code{"FLCI"}
(fixed-length two-sided confidence intervals)}
}
\description{
Computes the estimator and confidence intervals (CIs) for the CATT. The tuning parameter is chosen to optimize \code{opt.criterion} criterion.
}
\examples{
Ahalf <- diag(c(0.15, 0.6, 2.5, 2.5, 2.5, 0.5, 0.5, 0.1, 0.1))
D0 <- distMat(NSWexper[, 2:10], Ahalf, method="manhattan", NSWexper$treated)
## Distance matrix for variance estimation
DM <- distMat(NSWexper[, 2:10], Ahalf, method="manhattan")
sigma2 <- nnvar(DM, NSWexper$treated, NSWexper$re78, J=3)
## Compute homotopy/solution path, and the class of optimal estimators based
## on the solution path
op <- ATTOptPath(ATTh(D0, maxiter=200)$res, NSWexper$re78, NSWexper$treated)
ATTOptEstimate(op, mean(sigma2), C=1, sigma2final=sigma2, opt.criterion="RMSE")
ATTOptEstimate(op, mean(sigma2), C=1, sigma2final=sigma2, opt.criterion="FLCI")
}
