% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTOptEstimate}
\alias{ATTOptEstimate}
\title{build optimal estimator given a solution path}
\usage{
ATTOptEstimate(res, ep, d, y, C = 1, sigma2, sigma2final = sigma2,
  alpha = 0.05, beta = 0.8, opt.criterion = "RMSE", UpdateC = TRUE)
}
\arguments{
\item{res}{output of \code{ATTh}}

\item{ep}{output of \code{ATTEstimatePath}}

\item{d}{vector of treatment indicators}

\item{y}{outcome vector}

\item{C}{smoothness constant}

\item{sigma2}{vector of variances}

\item{sigma2final}{vector of variances for final variance.}

\item{alpha}{CI coverage}

\item{beta}{quantile of excess length}

\item{opt.criterion}{One of \code{"RMSE"}, \code{"OCI"},  \code{"FLCI"}}

\item{UpdateC}{Update C that's assumed in \code{ep}?}
}
\description{
build optimal estimator given a solution path
}
