% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTOptPath}
\alias{ATTOptPath}
\title{Compute the class of optimal linear estimators for the CATT}
\usage{
ATTOptPath(res, y, d)
}
\arguments{
\item{res}{The \code{res} element of the output of \code{ATTh} on which to
base the estimate}

\item{y}{outcome vector with length \code{n}}

\item{d}{Vector of treatment indicators with length \code{n}}
}
\description{
Computes the class of optimal linear estimators that minimize variance
subject to a bound on bias, and the optimal weights. The output of this
function is used by \code{\link{ATTOptEstimate}} for optimal estimation and
inference on the CATT.
}
