% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{ATTOptPath}
\alias{ATTOptPath}
\title{build optimal estimator given a solution path}
\usage{
ATTOptPath(res, y, d, sigma2, C = 1, alpha = 0.05, beta = 0.8)
}
\arguments{
\item{res}{output of \code{ATTh}}

\item{y}{outcome vector with length \code{n}}

\item{d}{Vector of treatment indicators with length \code{n}}

\item{sigma2}{vector of variance estimates with length \code{n}}

\item{C}{Lipschitz smoothness constant}

\item{alpha}{Level of confidence interval, \code{1-alpha}.}

\item{beta}{The quantile \code{beta} of excecss length for
determining performance of one-sided CIs.}
}
\description{
build optimal estimator given a solution path
}
