% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{distMat}
\alias{distMat}
\title{Compute matrix of distances between treated and untreated observations}
\usage{
distMat(x0, x1, Ahalf = diag(ncol(as.matrix(x1))), method = "euclidean")
}
\arguments{
\item{x0, x1}{Design matrices for treated and control variables with
dimensions \code{[n0 p]} and \code{n1 p}}

\item{Ahalf}{\eqn{A^{1/2}} weight matrix}

\item{method}{the distance measure to be used. This must be one of
    \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"},
    \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
    Any unambiguous substring can be given.}
}
\description{
Compute matrix of distances between treated and untreated observations
}
