% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distMat}
\alias{distMat}
\title{Matrix of distances between observations}
\usage{
distMat(X, Ahalf = diag(NCOL(X)), method = "euclidean", d = NULL, p = 2)
}
\arguments{
\item{X}{Design matrix of covariates with dimension \code{n} by \code{p}, or
else a vector of length \code{n} if there is a single covariate.}

\item{Ahalf}{Weight matrix with dimension \code{p} by \code{p} so that the
distances are computed between \code{Ahalf \%*\%  X[i, ]}.}

\item{method}{the distance measure to be used.  This must be one of
    \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"},
    \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
    Any unambiguous substring can be given.}

\item{d}{Vector of treatment indicators with length \code{n}. If supplied,
return the \code{n1} by \code{n0} submatrix corresponding to distances
between treated and untreated observations. Otherwise return the full
\code{n} by \code{n} matrix}

\item{p}{The power of the Minkowski distance.}
}
\value{
Matrix of distances with dimension \code{n} by \code{n} or else
    \code{n1} by \code{n0}
}
\description{
Compute a matrix of distances between \code{n} observations using the
distance measure in \code{method}.
}
\examples{
## 4 units, unit 1 and 3 are treated.
distMat(X=c(1, 2, 3, 4), d=c(TRUE, FALSE, TRUE, FALSE))
}
