% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{distMat}
\alias{distMat}
\title{Matrix of distances between observations}
\usage{
distMat(X, d, Ahalf = diag(NCOL(X)), method = "euclidean",
  FullMatrix = FALSE)
}
\arguments{
\item{X}{Design matrix for control variables with dimension \code{[n p]} (if
\eqn{p=1} and \code{X} is a vector, then it is converted to a matrix)}

\item{d}{Vector of treatment indicators}

\item{Ahalf}{\eqn{A^{1/2}} weight matrix so that the distances are computed
between \eqn{Ax_{0,j}} and \eqn{Ax_{1,i}}.}

\item{method}{the distance measure to be used. This must be one of
    \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"},
    \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
    Any unambiguous substring can be given.}

\item{FullMatrix}{Indicator for whether to return the full
\code{n0+n1}-by-\code{n0+n1} distance matrix, or only the
\code{n1}-by-\code{n0} distance matrix between treated and untreated
observations}
}
\value{
\code{[n1 n0]} or \code{[n n]} matrix of distances.
}
\description{
Compute and return the distance matrix using the distance measure in
\code{method}.
}
