% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_estimate.R
\name{nnvar}
\alias{nnvar}
\title{Nearest-neighbor variance estimator}
\usage{
nnvar(DM, d, y, J = 3)
}
\arguments{
\item{DM}{distance matrix}

\item{d}{treatment status}

\item{y}{outcome}

\item{J}{number of nearest neighbors}
}
\description{
Nearest-neighbor variance estimator
}
