% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nnvar}
\alias{nnvar}
\title{Nearest-neighbor variance estimator}
\usage{
nnvar(DM, d, y, J = 3, tol = 0)
}
\arguments{
\item{DM}{distance matrix with dimension \code{[n n]}.}

\item{d}{vector of treatment status indicators with length \code{n}}

\item{y}{outcome vector with length \code{n}}

\item{J}{number of nearest neighbors to average over}

\item{tol}{numerical tolerance for determining neighbors}
}
\description{
Calculates an \code{n}-vector sigma2 of estimates of the variance of \code{y}
using a nearest-neighbor esitmator among observations with the same treatment
status \code{d}.
}
